@@ -13,29 +13,30 @@
     <description>RuoYi-Vue-Plus多租户管理系统</description>
 
     <properties>
-        <revision>5.1.2</revision>
-        <spring-boot.version>3.1.7</spring-boot.version>
+        <revision>5.2.0-SNAPSHOT</revision>
+        <spring-boot.version>3.2.4</spring-boot.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <java.version>17</java.version>
         <spring-boot.mybatis>3.0.3</spring-boot.mybatis>
-        <springdoc.version>2.2.0</springdoc.version>
+        <springdoc.version>2.4.0</springdoc.version>
         <therapi-javadoc.version>0.15.0</therapi-javadoc.version>
         <poi.version>5.2.3</poi.version>
         <easyexcel.version>3.3.3</easyexcel.version>
         <velocity.version>2.3</velocity.version>
         <satoken.version>1.37.0</satoken.version>
-        <mybatis-plus.version>3.5.4</mybatis-plus.version>
+        <mybatis-plus.version>3.5.5</mybatis-plus.version>
         <p6spy.version>3.9.1</p6spy.version>
-        <hutool.version>5.8.22</hutool.version>
+        <hutool.version>5.8.26</hutool.version>
         <okhttp.version>4.10.0</okhttp.version>
-        <spring-boot-admin.version>3.1.8</spring-boot-admin.version>
-        <redisson.version>3.24.3</redisson.version>
-        <lock4j.version>2.2.5</lock4j.version>
-        <dynamic-ds.version>4.2.0</dynamic-ds.version>
+        <spring-boot-admin.version>3.2.3</spring-boot-admin.version>
+        <redisson.version>3.27.2</redisson.version>
+        <lock4j.version>2.2.7</lock4j.version>
+        <dynamic-ds.version>4.3.0</dynamic-ds.version>
         <alibaba-ttl.version>2.14.4</alibaba-ttl.version>
         <powerjob.version>4.3.6</powerjob.version>
-        <mapstruct-plus.version>1.3.5</mapstruct-plus.version>
+        <easyretry.version>3.2.0</easyretry.version>
+        <mapstruct-plus.version>1.3.6</mapstruct-plus.version>
         <mapstruct-plus.lombok.version>0.2.0</mapstruct-plus.lombok.version>
         <lombok.version>1.18.30</lombok.version>
         <bouncycastle.version>1.76</bouncycastle.version>
