@@ -1,8 +1,8 @@
 package org.dromara.common.job.config;
 
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import tech.powerjob.worker.PowerJobWorker;
 
@@ -11,7 +11,7 @@ import tech.powerjob.worker.PowerJobWorker;
  * @author yhan219
  * @since 2023/6/2
  */
-@Configuration
+@AutoConfiguration
 @ConditionalOnBean(PowerJobWorker.class)
 @ConditionalOnProperty(prefix = "powerjob.worker", name = "enabled", havingValue = "true")
 @EnableScheduling
