@@ -1,13 +1,10 @@
 package org.dromara.common.mybatis.interceptor;
 
-import cn.hutool.core.collection.ConcurrentHashSet;
-import cn.hutool.core.util.ArrayUtil;
 import com.baomidou.mybatisplus.core.plugins.InterceptorIgnoreHelper;
 import com.baomidou.mybatisplus.core.toolkit.PluginUtils;
 import com.baomidou.mybatisplus.extension.parser.JsqlParserSupport;
 import com.baomidou.mybatisplus.extension.plugins.inner.InnerInterceptor;
-import org.dromara.common.mybatis.annotation.DataColumn;
-import org.dromara.common.mybatis.handler.PlusDataPermissionHandler;
+import lombok.extern.slf4j.Slf4j;
 import net.sf.jsqlparser.expression.Expression;
 import net.sf.jsqlparser.statement.delete.Delete;
 import net.sf.jsqlparser.statement.select.PlainSelect;
@@ -22,11 +19,11 @@ import org.apache.ibatis.mapping.MappedStatement;
 import org.apache.ibatis.mapping.SqlCommandType;
 import org.apache.ibatis.session.ResultHandler;
 import org.apache.ibatis.session.RowBounds;
+import org.dromara.common.mybatis.handler.PlusDataPermissionHandler;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.List;
-import java.util.Set;
 
 /**
  * 数据权限拦截器
@@ -34,13 +31,14 @@ import java.util.Set;
  * @author Lion Li
  * @version 3.5.0
  */
+@Slf4j
 public class PlusDataPermissionInterceptor extends JsqlParserSupport implements InnerInterceptor {
 
-    private final PlusDataPermissionHandler dataPermissionHandler = new PlusDataPermissionHandler();
-    /**
-     * 无效注解方法缓存用于快速返回
-     */
-    private final Set<String> invalidCacheSet = new ConcurrentHashSet<>();
+    private final PlusDataPermissionHandler dataPermissionHandler;
+
+    public PlusDataPermissionInterceptor(String mapperPackage) {
+        this.dataPermissionHandler = new PlusDataPermissionHandler(mapperPackage);
+    }
 
     @Override
     public void beforeQuery(Executor executor, MappedStatement ms, Object parameter, RowBounds rowBounds, ResultHandler resultHandler, BoundSql boundSql) throws SQLException {
@@ -49,12 +47,7 @@ public class PlusDataPermissionInterceptor extends JsqlParserSupport implements 
             return;
         }
         // 检查是否无效 无数据权限注解
-        if (invalidCacheSet.contains(ms.getId())) {
-            return;
-        }
-        DataColumn[] dataColumns = dataPermissionHandler.findAnnotation(ms.getId());
-        if (ArrayUtil.isEmpty(dataColumns)) {
-            invalidCacheSet.add(ms.getId());
+        if (dataPermissionHandler.invalid(ms.getId())) {
             return;
         }
         // 解析 sql 分配对应方法
@@ -72,12 +65,7 @@ public class PlusDataPermissionInterceptor extends JsqlParserSupport implements 
                 return;
             }
             // 检查是否无效 无数据权限注解
-            if (invalidCacheSet.contains(ms.getId())) {
-                return;
-            }
-            DataColumn[] dataColumns = dataPermissionHandler.findAnnotation(ms.getId());
-            if (ArrayUtil.isEmpty(dataColumns)) {
-                invalidCacheSet.add(ms.getId());
+            if (dataPermissionHandler.invalid(ms.getId())) {
                 return;
             }
             PluginUtils.MPBoundSql mpBs = mpSh.mPBoundSql();
