@@ -14,19 +14,11 @@ import java.util.List;
 public interface SysPostMapper extends BaseMapperPlus<SysPost, SysPostVo> {
 
     /**
-     * 根据用户ID获取岗位选择框列表
-     *
-     * @param userId 用户ID
-     * @return 选中岗位ID列表
-     */
-    List<Long> selectPostListByUserId(Long userId);
-
-    /**
      * 查询用户所属岗位组
      *
-     * @param userName 用户名
+     * @param userId 用户ID
      * @return 结果
      */
-    List<SysPostVo> selectPostsByUserName(String userName);
+    List<SysPostVo> selectPostsByUserId(Long userId);
 
 }
