@@ -31,12 +31,11 @@ public interface ISysOssConfigService {
      */
     TableDataInfo<SysOssConfigVo> queryPageList(SysOssConfigBo bo, PageQuery pageQuery);
 
-
     /**
      * 根据新增业务对象插入对象存储配置
      *
      * @param bo 对象存储配置新增业务对象
-     * @return
+     * @return 结果
      */
     Boolean insertByBo(SysOssConfigBo bo);
 
@@ -44,7 +43,7 @@ public interface ISysOssConfigService {
      * 根据编辑业务对象修改对象存储配置
      *
      * @param bo 对象存储配置编辑业务对象
-     * @return
+     * @return 结果
      */
     Boolean updateByBo(SysOssConfigBo bo);
 
@@ -53,7 +52,7 @@ public interface ISysOssConfigService {
      *
      * @param ids     主键集合
      * @param isValid 是否校验,true-删除前校验,false-不校验
-     * @return
+     * @return 结果
      */
     Boolean deleteWithValidByIds(Collection<Long> ids, Boolean isValid);
 
