@@ -6,6 +6,8 @@ import cn.hutool.core.util.ObjectUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
 import org.dromara.common.core.constant.CacheNames;
 import org.dromara.common.core.exception.ServiceException;
 import org.dromara.common.core.service.OssService;
@@ -25,8 +27,6 @@ import org.dromara.system.domain.bo.SysOssBo;
 import org.dromara.system.domain.vo.SysOssVo;
 import org.dromara.system.mapper.SysOssMapper;
 import org.dromara.system.service.ISysOssService;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.http.MediaType;
@@ -36,7 +36,10 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 
 /**
  * 文件上传 服务层实现
@@ -49,6 +52,13 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
 
     private final SysOssMapper baseMapper;
 
+    /**
+     * 查询OSS对象存储列表
+     *
+     * @param bo        OSS对象存储分页查询对象
+     * @param pageQuery 分页查询实体类
+     * @return 结果
+     */
     @Override
     public TableDataInfo<SysOssVo> queryPageList(SysOssBo bo, PageQuery pageQuery) {
         LambdaQueryWrapper<SysOss> lqw = buildQueryWrapper(bo);
@@ -58,6 +68,12 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         return TableDataInfo.build(result);
     }
 
+    /**
+     * 根据一组 ossIds 获取对应的 SysOssVo 列表
+     *
+     * @param ossIds 一组文件在数据库中的唯一标识集合
+     * @return 包含 SysOssVo 对象的列表
+     */
     @Override
     public List<SysOssVo> listByIds(Collection<Long> ossIds) {
         List<SysOssVo> list = new ArrayList<>();
@@ -75,6 +91,12 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         return list;
     }
 
+    /**
+     * 根据一组 ossIds 获取对应文件的 URL 列表
+     *
+     * @param ossIds 以逗号分隔的 ossId 字符串
+     * @return 以逗号分隔的文件 URL 字符串
+     */
     @Override
     public String selectUrlByIds(String ossIds) {
         List<String> list = new ArrayList<>();
@@ -107,12 +129,25 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         return lqw;
     }
 
+    /**
+     * 根据 ossId 从缓存或数据库中获取 SysOssVo 对象
+     *
+     * @param ossId 文件在数据库中的唯一标识
+     * @return SysOssVo 对象，包含文件信息
+     */
     @Cacheable(cacheNames = CacheNames.SYS_OSS, key = "#ossId")
     @Override
     public SysOssVo getById(Long ossId) {
         return baseMapper.selectVoById(ossId);
     }
 
+
+    /**
+     * 文件下载方法，支持一次性下载完整文件
+     *
+     * @param ossId    OSS对象ID
+     * @param response HttpServletResponse对象，用于设置响应头和向客户端发送文件内容
+     */
     @Override
     public void download(Long ossId, HttpServletResponse response) throws IOException {
         SysOssVo sysOss = SpringUtils.getAopProxy(this).getById(ossId);
@@ -122,7 +157,7 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         FileUtils.setAttachmentResponseHeader(response, sysOss.getOriginalName());
         response.setContentType(MediaType.APPLICATION_OCTET_STREAM_VALUE + "; charset=UTF-8");
         OssClient storage = OssFactory.instance(sysOss.getService());
-        try(InputStream inputStream = storage.getObjectContent(sysOss.getUrl())) {
+        try (InputStream inputStream = storage.getObjectContent(sysOss.getUrl())) {
             int available = inputStream.available();
             IoUtil.copy(inputStream, response.getOutputStream(), available);
             response.setContentLength(available);
@@ -131,6 +166,13 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         }
     }
 
+    /**
+     * 上传 MultipartFile 到对象存储服务，并保存文件信息到数据库
+     *
+     * @param file 要上传的 MultipartFile 对象
+     * @return 上传成功后的 SysOssVo 对象，包含文件信息
+     * @throws ServiceException 如果上传过程中发生异常，则抛出 ServiceException 异常
+     */
     @Override
     public SysOssVo upload(MultipartFile file) {
         String originalfileName = file.getOriginalFilename();
@@ -138,7 +180,7 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         OssClient storage = OssFactory.instance();
         UploadResult uploadResult;
         try {
-            uploadResult = storage.uploadSuffix(file.getBytes(), suffix, file.getContentType());
+            uploadResult = storage.uploadSuffix(file.getBytes(), suffix);
         } catch (IOException e) {
             throw new ServiceException(e.getMessage());
         }
@@ -146,6 +188,12 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         return buildResultEntity(originalfileName, suffix, storage.getConfigKey(), uploadResult);
     }
 
+    /**
+     * 上传文件到对象存储服务，并保存文件信息到数据库
+     *
+     * @param file 要上传的文件对象
+     * @return 上传成功后的 SysOssVo 对象，包含文件信息
+     */
     @Override
     public SysOssVo upload(File file) {
         String originalfileName = file.getName();
@@ -169,6 +217,13 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
         return this.matchingUrl(sysOssVo);
     }
 
+    /**
+     * 删除OSS对象存储
+     *
+     * @param ids     OSS对象ID串
+     * @param isValid 判断是否需要校验
+     * @return 结果
+     */
     @Override
     public Boolean deleteWithValidByIds(Collection<Long> ids, Boolean isValid) {
         if (isValid) {
@@ -183,7 +238,7 @@ public class SysOssServiceImpl implements ISysOssService, OssService {
     }
 
     /**
-     * 匹配Url
+     * 桶类型为 private 的URL 修改为临时URL时长为120s
      *
      * @param oss OSS对象
      * @return oss 匹配Url的OSS对象
